Repo -> JPA->CRUD

+++ JPA_View/src/main/java/com/jpa/entity/User.java
@@ -7,7 +7,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long Id;
+    private int Id;
     private String name;
 
     private String email;
@@ -20,7 +20,7 @@ public class User {
     public User() {
     }
 
-    public User(Long id, String name, String email, String address, String status) {
+    public User(int id, String name, String email, String address, String status) {
         Id = id;
         this.name = name;
         this.email = email;
@@ -29,11 +29,11 @@ public class User {
     }
 
 
-    public Long getId() {
+    public int getId() {
         return Id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         Id = id;
     }
